+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+- 2025-01-14: updated packages and fix to messaging in index command.
 - 2025-01-11:
   - updated test packages.
   - added notification to indexing command in CLI tool. Please notice that currently this uses [MailJet](https://www.mailjet.com) and it requires you to save your API keys into environment variables (named `MAILJET_API_KEY_PUBLIC` and `MAILJET_API_KEY_PRIVATE`).